develop 브랜치 충돌 해결 2
ExController 클래스와 GlobalExceptionHandler 클래스가 중복되어 합침
