+++ src/main/java/oneseoktwojo/ohtalkhae/global/exception/GlobalExceptionHandler.java
@@ -1,17 +1,27 @@
 package oneseoktwojo.ohtalkhae.global.exception;
 
+import lombok.extern.slf4j.Slf4j;
 import oneseoktwojo.ohtalkhae.global.dto.ApiResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.http.converter.HttpMessageNotReadableException;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
+import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
 
+@Slf4j
 @RestControllerAdvice
 public class GlobalExceptionHandler {
+
+    @ExceptionHandler(HttpMessageNotReadableException.class)
+    public ApiResponse<?> handleHttpMessageNotReadable(HttpMessageNotReadableException ex) {
+        log.debug(ex.getMessage());
+        return ApiResponse.error(400, "Invalid Request Body");
+    }
 
     @ExceptionHandler(IllegalArgumentException.class)
     public ResponseEntity<ApiResponse<?>> handleIllegalArgumentException(IllegalArgumentException ex) {
@@ -21,6 +31,11 @@ public class GlobalExceptionHandler {
     @ExceptionHandler(IllegalStateException.class)
     public ResponseEntity<ApiResponse<?>> handleIllegalStateException(IllegalStateException ex) {
         return ResponseEntity.status(HttpStatus.CONFLICT).body(ApiResponse.error(409, ex.getMessage()));
+    }
+
+    @ExceptionHandler(IOException.class)
+    public ApiResponse<?> handleIOException(IOException ex) {
+        return ApiResponse.error(HttpStatus.INTERNAL_SERVER_ERROR.value(), "프로필 사진 업로드에 실패했습니다 : " + ex.getMessage());
     }
 
     @ExceptionHandler(MethodArgumentNotValidException.class)
